backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart==0.0.6
Pillow>=10.3.0
SQLAlchemy>=2.0.30
chromadb
langchain-ollama
langchain-community
PyMySQL==1.1.0
cryptography==41.0.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.28.1